backend/requirements.txt
# pip install -r requirements.txt

django==2.2.1
psycopg2==2.7.6.1
djangorestframework==3.9.3
django-cors-headers==2.0.0

